+++ Changelog.md
@@ -8,6 +8,7 @@ Compiler Features:
 * JSON-AST: Added selector field for errors and events.
  * Peephole Optimizer: Optimize comparisons in front of conditional jumps and conditional jumps across a single unconditional jump.
  * Yul Optimizer: Remove ``sstore`` and ``mstore`` operations that are never read from.
+ * LSP: Implements goto-definition.
 
 Bugfixes:
  * Yul IR Code Generation: Optimize embedded creation code with correct settings. This fixes potential mismatches between the constructor code of a contract compiled in isolation and the bytecode in ``type(C).creationCode``, resp. the bytecode used for ``new C(...)``.